chore: moved mongo related code to single file - moved mongo related code to a single file that was being used - removed empty files
