+++ forum/models/model_utils.py
@@ -1,8 +1,8 @@
 """Model util function for db operations."""
 
-from typing import Any, Dict, Union
+from typing import Any, Dict, Tuple, Union
 
-from forum.models import Comment, Contents, Users
+from forum.models import Comment, CommentThread, Contents, Users
 
 
 def flag_as_abuse(
@@ -45,7 +45,9 @@ def flag_as_abuse(
     return Contents().get(entity["_id"])
 
 
-def un_flag_as_abuse(user: Dict[str, Any], entity: Dict[str, Any]) -> Union[Dict[str, Any], None]:
+def un_flag_as_abuse(
+    user: Dict[str, Any], entity: Dict[str, Any]
+) -> Union[Dict[str, Any], None]:
     """
     Unflag an entity as abuse.
 
@@ -203,3 +205,93 @@ def get_comments_count(thread_id: str) -> int:
     """
     comments = list(Comment().list(comment_thread_id=thread_id))
     return len(comments) if comments else 0
+
+
+def validate_thread_and_user(
+    user_id: str, thread_id: str
+) -> Tuple[Dict[str, Any], Dict[str, Any]]:
+    """
+    Validate thread and user.
+
+    Arguments:
+        user_id (str): The ID of the user making the request.
+        thread_id (str): The ID of the thread.
+
+    Returns:
+        Tuple[Dict[str, Any], Dict[str, Any]]: A tuple containing the user and thread data.
+
+    Raises:
+        ValueError: If the thread or user is not found.
+    """
+    thread = CommentThread().get(thread_id)
+    user = Users().get(user_id)
+    if not (thread and user):
+        raise ValueError("User / Thread doesn't exist")
+
+    return user, thread
+
+
+def pin_unpin_thread(thread_id: str, action: str) -> None:
+    """
+    Pin or unpin the thread based on action parameter.
+
+    Arguments:
+        thread_id (str): The ID of the thread to pin/unpin.
+        action (str): The action to perform ("pin" or "unpin").
+    """
+    CommentThread().update(thread_id, pinned=action == "pin")
+
+
+def get_pinned_unpinned_thread_serialized_data(
+    user: Dict[str, Any], thread_id: str, serializer_class: Any
+) -> Dict[str, Any]:
+    """
+    Return serialized data of pinned or unpinned thread.
+
+    Arguments:
+        user (Dict[str, Any]): The user who requested the action.
+        thread_id (str): The ID of the thread to pin/unpin.
+
+    Returns:
+        Dict[str, Any]: The serialized data of the pinned/unpinned thread.
+
+    Raises:
+        ValueError: If the serialization is not valid.
+    """
+    updated_thread = CommentThread().get(thread_id)
+    context = {
+        "user_id": user["_id"],
+        "username": user["username"],
+        "type": "thread",
+        "id": thread_id,
+    }
+    if updated_thread is not None:
+        context = {**context, **updated_thread}
+    serializer = serializer_class(data=context)
+    if not serializer.is_valid():
+        raise ValueError(serializer.errors)
+
+    return serializer.data
+
+
+def handle_pin_unpin_thread_request(
+    user_id: str, thread_id: str, action: str, serializer_class: Any
+) -> Dict[str, Any]:
+    """
+    Catches pin/unpin thread request.
+
+    - validates thread and user.
+    - pin or unpin the thread based on action parameter.
+    - return serialized data of thread.
+
+    Arguments:
+        user_id (str): The ID of the user making the request.
+        thread_id (str): The ID of the thread to pin/unpin.
+        action (str): The action to perform ("pin" or "unpin").
+
+    Returns:
+        Dict[str, Any]: The serialized data of the pinned/unpinned thread.
+    """
+    user, _ = validate_thread_and_user(user_id, thread_id)
+    pin_unpin_thread(thread_id, action)
+    return get_pinned_unpinned_thread_serialized_data(user, thread_id, serializer_class)
